Move MacroDependsOn and Docs (circular reference)
